% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{basic_analysis}
\alias{basic_analysis}
\title{Basic Analysis of Fixations}
\usage{
basic_analysis(data, eye_tracker = "SMI", object, participant)
}
\arguments{
\item{data}{data frame}

\item{eye_tracker}{the type of used eye-tracker ("SMI"/"Tobii"). Default value "SMI". (character)}

\item{object}{the name of a stimuli (character)}

\item{participant}{the name of a participant or participants (character or vector of characters)}
}
\value{
Data frame with the following variables:
\itemize{
  \item NoF - Number of Fixations
  \item TTFF - Time to First Fixation [ms] - NA for Tobii data
  \item FFD - First Fixation Duration [ms]
  \item TD - Total Duration [ms]
  \item AFD - Average Fixation Duration [ms]
}
}
\description{
It offers a basic analysis for the initial conception of the number and duration of fixations.
The data from SMI eye-trackers must contain the following variables:
\itemize{
    \item Stimulus (Presented Media name)
    \item Participant (Participant name)
    \item Index	(Eye movement type index)
    \item Event Start Trial Time [ms]
    \item Event End Trial Time [ms]
    \item Event Duration [ms] (Gaze event duration)
    \item Fixation Position X [px]	(Fixation point X)
    \item Fixation Position Y [px] (Fixation point Y)
    \item AOI Name (AOI hit)
    \item Eye movement type - only for Tobii
}
You will be able to find the relevant variables for Tobii eye trackers within parenthesis.
}
\examples{
basic_analysis(data_SMI, "SMI", "09-M1-CX-SI-VE.jpg", "P16")
basic_analysis(data_Tobii, "Tobii", "10.jpg", "Participant19")

}
