% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coeffK_single.R
\name{coeffK_single}
\alias{coeffK_single}
\title{Ambient/Focal Attention - Coefficient K - single}
\usage{
coeffK_single(
  data,
  eye_tracker = "SMI",
  object,
  participant,
  distance,
  lim = c(-10, 10),
  point_col = "steelblue",
  point_size = 3,
  title_size = 13,
  x_size = 10,
  x_angle = 0,
  x_labsize = 10,
  y_size = 10,
  y_angle = 0,
  y_labsize = 10
)
}
\arguments{
\item{data}{data frame}

\item{eye_tracker}{the type of used eye-tracker ("SMI"/"Tobii"). Default value "SMI". (character)}

\item{object}{the name of a stimuli (character)}

\item{participant}{the name of a participant(character)}

\item{distance}{distance from screen and eyes [cm]}

\item{lim}{y axis scale (numeric vector)}

\item{point_col}{colour of points. Default value "steelblue".}

\item{point_size}{size of points}

\item{title_size}{the size of title}

\item{x_size}{x lab ticks font size}

\item{x_angle}{x lab text angle}

\item{x_labsize}{x lab font size}

\item{y_size}{y lab ticks font size}

\item{y_angle}{y lab text angle}

\item{y_labsize}{y lab font size}
}
\value{
List - data frame (K - elements of coefficient K) and plot (K_plot - visualization of elements K_i, title includes final coefficient K)
}
\description{
Calculate and plot elements of Coefficient K. Coefficient K helps distinguish focal and ambient attention.
}
\details{
As the attention and saccades are interconnected, it is theorised that the duration of fixations and amplitide of saccades reflects the different types of attention. In order to distinguish types of attantion, a cooefficient K has been intoduced as a metric to characterise ambient and focal attention when conduciting cartographic tasks. Ambient attention is typically characterised by relatively short fixations, which are followed by long saccades. Conversely, focal attention is defined by long fixations which are followed by short saccades. K > 0 denotes focal attention, whereas K < 0 denotes ambient attention. If our participant's attention transitions from ambient to focal, then K can indicate the end of visual searching, either triggered by boredom or making a decision according to the task's instructions.

The data from SMI eye-trackers must contain the following variables:
\itemize{
    \item Stimulus (Presented Media name)
    \item Participant (Participant name)
    \item Index	(Eye movement type index)
    \item Event Start Trial Time [ms]
    \item Event End Trial Time [ms]
    \item Event Duration [ms] (Gaze event duration)
    \item Fixation Position X [px]	(Fixation point X)
    \item Fixation Position Y [px] (Fixation point Y)
    \item AOI Name (AOI hit)
    \item Eye movement type - only for Tobii
}
You will be able to find the relevant variables for Tobii eye trackers within parenthesis.
}
\examples{
coeffK_single(data_SMI, "SMI", "09-M1-CX-SI-VE.jpg", "P16", 60, c(-5, 5))
coeffK_single(data_Tobii, "Tobii", "10.jpg", "Participant19", 60)

}
