% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_SMI.R
\docType{data}
\name{data_SMI}
\alias{data_SMI}
\title{Data SMI}
\format{
Data frame with 25 rows and 9 variables:
\itemize{
    \item Stimulus - the name of a stimuli
    \item Participant - the name of a participant
    \item Index - Index of fixations
    \item Event.Start.Trial.Time..ms - Start of fixation [ms]
    \item Event.End.Trial.Time..ms - End of fixation [ms]
    \item Event.Duration..ms - Duration of fixations [ms]
    \item Fixation.Position.X..px - X-position of fixation [px]
    \item Fixation.Position.Y..px - Y-Position of fixation [px]
    \item AOI.Name - Area of interest of each fixation
}
}
\usage{
data_SMI
}
\description{
Eye-tracking data from eye-tracker SMI.
}
\keyword{datasets}
