% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_Tobii.R
\docType{data}
\name{data_Tobii}
\alias{data_Tobii}
\title{Data Tobii}
\format{
A data frame with 15 rows and 9 variables:
\itemize{
    \item Participant.name - the name of a participant
    \item Presented.Media.name - the name of stimuli
    \item Eye.movement.type - type of eye movement (fixation)
    \item Gaze.event.duration - Duration of fixation
    \item Eye.movement.type.index - Index of fixation
    \item Fixation.point.X - X-position of fixation [px]
    \item Fixation.point.Y - Y-position of fixation [px]
    \item AOI.hit..10...right - 1/0 hit right area of interest
    \item AOI.hit..10...left - 1/0 hit left area of interest
}
}
\usage{
data_Tobii
}
\description{
Eye-tracking data from eye-tracker Tobii.
}
\keyword{datasets}
