% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{measures}
\alias{measures}
\title{Recurrence Quantification Analysis Measures}
\usage{
measures(
  data,
  eye_tracker = "SMI",
  object,
  participant,
  method = "Fixed distance",
  r = 60,
  size = c(1920, 1200),
  L = 2,
  time = FALSE
)
}
\arguments{
\item{data}{data frame}

\item{eye_tracker}{the type of used eye-tracker ("SMI"/"Tobii"). Default value "SMI". (character)}

\item{object}{the name of a stimuli (character)}

\item{participant}{the name of a participant or participants (character or vector of characters)}

\item{method}{the name of used method ("fixed distance", "fixed grid", "AOI"). Default value "fixed distance". (character)}

\item{r}{radius of circle (fixed distance), size of square (fixed grid)' @param size dimension of stimulus (c(width, height)). Default value c(1920, 1200).}

\item{size}{size of stimulus (vector - width, height). Default c(1920, 1200).}

\item{L}{length of diagonal, horizontal and vertical lines. Default value set to 2.}

\item{time}{involve duration of fixations (TRUE/FALSE). Default value FALSE.}
}
\value{
data frame that contains of
\itemize{
  \item R - the sum of recurrences in the upper triangle of the recurrence plot/matrix.
  \item REC - Recurrence (\%) - recurence measure represents, for a sequence of N fixations, the percentage of recurrent fixations (i.e., how often observers refixate previously fixated image positions).
  \item DET - Determinism (\%) - the determinism measures the proportion of recurrent points forming diagonal lines and represents repeating gaze patterns in the recurrence plot. The minimum line length of diagonal line elements was set to L = 2. The length of the diagonal line element reflects the number of fixations making up the repeated scan path.
  \item LAM - Laminarity (\%) - he laminary measure is determined by vertical and horizontal lines. Vertical lines represent areas that were fixated first in a single fixation and then rescanned in detail over consecutive fixations at a later time (e.g., several fixations later), and horizontal lines represent areas that were first scanned in detail and then refixated briefly later in time. Again, we set the minimum line lengths of vertical and horizontal lines to L = 2.
  \item CORM - Center of Recurrent Mass - this measure indicates approximately where in time most of the recurrent points are situated. Small corm values indicate that refixations tend to occur close in time, whereas large corm values indicate that refixations tend to occur widely separated in time.
}
}
\description{
Calculate RQA Measures.
}
\details{
Visualization of recurrent fixations can often predispose to subjective bias when evaluating a set of results. For that reason, we used recurrence quantification analysis measures, which allow us to quantify data displayed in the recurrence plot. Using RQA, we can compare different tasks or compare multiple participants.

The data from SMI eye-trackers must contain the following variables:
\itemize{
    \item Stimulus (Presented Media name)
    \item Participant (Participant name)
    \item Index	(Eye movement type index)
    \item Event Start Trial Time [ms]
    \item Event End Trial Time [ms]
    \item Event Duration [ms] (Gaze event duration)
    \item Fixation Position X [px]	(Fixation point X)
    \item Fixation Position Y [px] (Fixation point Y)
    \item AOI Name (AOI hit)
    \item Eye movement type - only for Tobii
}
You will be able to find the relevant variables for Tobii eye trackers within parenthesis.
}
\examples{
measures(data_SMI, "SMI", "09-M1-CX-SI-VE.jpg", "P16", "Fixed distance")
measures(data_Tobii, "Tobii", "10.jpg", "Participant19", "AOI", size = c(1920, 1080), time = TRUE)

}
