% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recurrence.R
\name{recurrence}
\alias{recurrence}
\title{Recurrence}
\usage{
recurrence(
  data,
  eye_tracker = "SMI",
  object,
  participant,
  method = "Fixed distance",
  r = 60,
  scaling = 100,
  size = c(1920, 1200),
  time = FALSE,
  col_man = FALSE,
  col_set = c(),
  point_size = 1.5,
  point_col = "darkblue",
  bubble_col = "steelblue",
  title_size = 15,
  legend_size = 5,
  x_labsize = 10,
  y_labsize = 10,
  x_size = 10,
  y_size = 10,
  breaks = 2,
  lightness = 0.2
)
}
\arguments{
\item{data}{data frame}

\item{eye_tracker}{the type of used eye-tracker ("SMI"/"Tobii"). Default value "SMI". (character)}

\item{object}{the name of a stimulus (character)}

\item{participant}{the name of a participant (character)}

\item{method}{the name of used method ("fixed distance", "fixed grid", "AOI"). Default value "fixed distance".}

\item{r}{radius of circle (fixed distance), size of square (fixed grid)}

\item{scaling}{size of bubble (representing relative time)}

\item{size}{size of stimulus (vector - width, height). Default c(1920, 1200)}

\item{time}{involve time TRUE/FALSE. Default value FALSE.}

\item{col_man}{for AOI methods. If you do not want to set colours of points manually (FALSE). If you want to set colours of points manually (TRUE). Default value is FALSE.}

\item{col_set}{vector of N colours (N - number of unique AOIs in dataset). If col_man = T.}

\item{point_size}{size of points}

\item{point_col}{colour of points (fixed distance and fixed grid)}

\item{bubble_col}{colour of bubble representing relative duration of fixation (fixed distance, fixed grid with time involved)}

\item{title_size}{title font size}

\item{legend_size}{legend font size}

\item{x_labsize}{x lab font size}

\item{y_labsize}{y lab font size}

\item{x_size}{x lab ticks font size}

\item{y_size}{y lab ticks font size}

\item{breaks}{breaks on x and y axis}

\item{lightness}{lightness of bubble in RP}
}
\value{
list Recurrence matrix, Recurrence plot
}
\description{
This function returns Recurrence matrix and Recurrence plot. Two fixations are considered to be recurrent if they are close together.
We use 3 different methods to identify recurrent fixations:
\itemize{
   \item fixed distance - two fixations \emph{A} and \emph{B} are recurrent if they are close to each other (i.e., if the Euclidean distance \emph{d(A, B) <= r} for a fixed radius \emph{r}).
   \item fixed grid - fixation map overlaid by a grid with element size of \emph{r} pixels . Two fixations \emph{A} and \emph{B} are considered recurrent if they land in the same grid element.
   \item Areas of Interest (AOI) - two fixations \emph{A} and \emph{B} are considered recurrent if they land in the same area of interest
}
}
\details{
Recurrence matrix is a square matrix, which includes 1 if the fixation is recurrent and 0 if it is not recurrent. Should a duration be included, the matrix will include the sum of durations instead of 1. Recurrence can be represented in a recurrence plot, which plots recurrences of a fixation sequence with itself over all possible time lags. If fixations i and j are recurrent (i.e., if r_ij = 1), then a dot is plotted at position i, j. If we are including durations as a variable, the size of a point on a plot is given by the sum of recurrence of both fixations t_i a t_j.

The data from SMI eye-trackers must contain the following variables:
\itemize{
    \item Stimulus (Presented Media name)
    \item Participant (Participant name)
    \item Index	(Eye movement type index)
    \item Event Start Trial Time [ms]
    \item Event End Trial Time [ms]
    \item Event Duration [ms] (Gaze event duration)
    \item Fixation Position X [px]	(Fixation point X)
    \item Fixation Position Y [px] (Fixation point Y)
    \item AOI Name (AOI hit)
    \item Eye movement type - only for Tobii
}
You will be able to find the relevant variables for Tobii eye trackers within parenthesis.
}
\examples{
recurrence(data_SMI, "SMI", "09-M1-CX-SI-VE.jpg", "P16", "Fixed distance", point_col = "blue", bubble_col = "steelblue", legend_size = 10, lightness = 0.2)$Rec_plot
recurrence(data_Tobii, "Tobii", "10.jpg", "Participant19", "AOI", size = c(1920, 1080),  time = FALSE)

}
