% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{visualization}
\alias{visualization}
\title{Visualization of fixation}
\usage{
visualization(
  data,
  eye_tracker = "SMI",
  object,
  participant,
  method = "Simple",
  fig = FALSE,
  img = NULL,
  img_AOI = NULL,
  scaling = 100,
  r = 60,
  size = c(1920, 1200),
  col_man = F,
  col_set = c(),
  point_size = 2,
  point_col = "steelblue",
  bubble_col = "steelblue",
  title_size = 15,
  x_labsize = 10,
  y_labsize = 10,
  legend_size = 10,
  text_size = 5,
  sac_size = 0.2,
  circle_size = 0.1,
  grid_size = 0.3,
  lightness = 0.4
)
}
\arguments{
\item{data}{data frame}

\item{eye_tracker}{the type of used eye-tracker ("SMI"/"Tobii"). Default value "SMI". (character)}

\item{object}{the name of a stimulus (character)}

\item{participant}{the name of a participant (character)}

\item{method}{visualization method - "Simple", "Simple with time", "Fixed distance", "Fixed grid", "AOI". Default "Simple"}

\item{fig}{if you want to put image of stimulus under plot (TRUE/FALSE). Default FALSE.}

\item{img}{image of stimulus (.jpg, .png). If fig = T and method is not with AOI.}

\item{img_AOI}{image of stimulus with AOI (.jpg, .png). If fig = T & method = "AOI" ("AOI with time").}

\item{scaling}{size of bubble}

\item{r}{radius of circle (fixed grid), size of square (fixed grid)}

\item{size}{size of stimulus (vector - width, height). Default c(1920, 1200)}

\item{col_man}{for AOI methods. If you do not want to set colours of points manually (FALSE). If you want to set colours of points manually (TRUE). Default value is FALSE.}

\item{col_set}{vector of N colours (N - number of unique AOIs in dataset). If col_man = T.}

\item{point_size}{size of points}

\item{point_col}{colour of points (for methods - simple, simple wt, fixed distance, fixed grid)}

\item{bubble_col}{colour of bubble representing relative duration of fixation (only for simple with time)}

\item{title_size}{title font size}

\item{x_labsize}{x lab font size}

\item{y_labsize}{y lab font size}

\item{legend_size}{legend font size (AOI methods)}

\item{text_size}{font size in plot}

\item{sac_size}{size of saccades}

\item{circle_size}{size of circles (fixed distance)}

\item{grid_size}{size of grid (fixed grid)}

\item{lightness}{lightness of bubbles representing relative duration of fixation (only for simple with time and AOI with time)}
}
\value{
Plot of scanpath according to visualization method of your choice.
}
\description{
This function offers visualization of the scanpath above the stimulus. You can choose a method of visualization. You have following options:
\itemize{
    \item Simple - visualization of scanpath
    \item Simple with time - visualization of scanpath with duration of fixations
    \item Fixed distance - visualization of scanpath, including a circle with a radius of your choice
    \item Fixed Grid - visualization of scanpath with a grid (you can choose the size of the dimensions)
    \item AOI - offers visualization of scanpath. The colours of fixations vary based on its placement within different AOIs
    \item AOI with time - offers same visualization as the AOI method but also includes duration of fixations
}
}
\details{
The data from SMI eye-trackers must contain the following variables:
\itemize{
    \item Stimulus (Presented Media name)
    \item Participant (Participant name)
    \item Index	(Eye movement type index)
    \item Event Start Trial Time [ms]
    \item Event End Trial Time [ms]
    \item Event Duration [ms] (Gaze event duration)
    \item Fixation Position X [px]	(Fixation point X)
    \item Fixation Position Y [px] (Fixation point Y)
    \item AOI Name (AOI hit)
    \item Eye movement type - only for Tobii
}
You will be able to find the relevant variables for Tobii eye trackers within parenthesis.
}
\examples{
visualization(data_SMI, "SMI","09-M1-CX-SI-VE.jpg", "P16", method = "Simple", point_size = 3, text_size = 8)
visualization(data_Tobii, "Tobii","10.jpg", "Participant19", method = "AOI", size = c(1920, 1080))
}
